refactor: thin admin settings page into subcomponents (1.0.10)
app/admin/settings/page.tsx drops from 691 to 253 lines, keeping
configs/users state, fetching, and all save handlers as a container.

Extracted:
- constants.ts: types, config keys, CONFIG_MAX_LENGTHS,
  SUSPICIOUS_PATTERNS/hasSuspiciousContent, CONFIG_SECTIONS,
  MODEL_FIELDS/MODEL_DEFAULTS
- components/: ModelConfigSection, AutomationSection,
  ConfigSectionsList, EditConfigModal

Inline anonymous onChange/onClick logic pulled into named handlers
(selectModel/resetModels/toggleAutoSync). Toast strings and reset
defaults preserved verbatim, including the existing quirk where
failover/summary reset to flash-lite. UI and behavior unchanged; no
API change. tsc, next build, and 25 tests all pass.
