+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "antnest-chatbot",
-  "version": "1.0.9",
+  "version": "1.0.10",
   "description": "",
   "main": "index.js",
   "scripts": {
